rag_quotes/requirements.txt
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
chromadb>=0.4.22
tqdm>=4.66.0 